langserve-app/pyproject.toml
[tool.poetry]
name = "my-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
rag-conversation = {path = "packages/rag-conversation", develop = true}
extraction-openai-functions = {path = "packages/extraction-openai-functions", develop = true}
openai-functions-agent = {path = "packages/openai-functions-agent", develop = true}
anthropic-iterative-search = {path = "packages/anthropic-iterative-search", develop = true}
gemini-functions-agent = {path = "packages/gemini-functions-agent", develop = true}
openai-functions-tool-retrieval-agent = {path = "packages/openai-functions-tool-retrieval-agent", develop = true}


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
